added edit note feature, fixed the reference issue after permanently deleting a note

+++ controllers/binController.js
@@ -1,4 +1,5 @@
 import Notes from "../models/Notes.model.js";
+import User from "../models/User.model.js";
 
 export const restoreFromBin = async (req, res) => {
   try {
@@ -18,21 +19,40 @@ export const deleteFromBin = async (req, res) => {
     const { noteId } = req.body;
     const userId = req.user.id;
 
-    await Notes.findOneAndDelete({ _id: noteId, user: userId, isDeleted: true });
+    const deletedNote = await Notes.findOneAndDelete({ _id: noteId, user: userId, isDeleted: true });
+
+    // If the note was successfully deleted, update the user's notes array
+    if (deletedNote) {
+      await User.findByIdAndUpdate(userId, {
+        $pull: { notes: noteId }, // Remove the note reference from the user's notes array
+      });
+    }
+
     res.redirect("/bin");
   } catch (error) {
-    console.error(err);
-    res.status(500).send(err.message);
+    console.error(error);
+    res.status(500).send(error.message);
   }
 };
 
 export const clearBin = async (req, res) => {
   try {
     const userId = req.user.id;
-    await Notes.deleteMany({ user: userId, isDeleted: true });
+
+    const deletedNotes = await Notes.find({ user: userId, isDeleted: true });
+
+    if (deletedNotes.length > 0) {
+      await Notes.deleteMany({ user: userId, isDeleted: true });
+
+      const noteIds = deletedNotes.map((note) => note._id);
+      await User.findByIdAndUpdate(userId, {
+        $pull: { notes: { $in: noteIds } },
+      });
+    }
+
     res.redirect("/bin");
   } catch (error) {
-    console.error(err);
-    res.status(500).send(err.message);
+    console.error(error);
+    res.status(500).send(error.message);
   }
 };
